+++ src/mongo/connectToMongoClient.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { MongoClient, MongoClientOptions } from 'mongodb';
-import { Links } from '../constants';
+import { emulatorPassword, Links } from '../constants';
 
 export async function connectToMongoClient(connectionString: string, appName: string): Promise<MongoClient> {
     // appname appears to be the correct equivalent to user-agent for mongo
@@ -15,20 +15,34 @@ export async function connectToMongoClient(connectionString: string, appName: st
         useNewUrlParser: true
     };
 
+    if (isCosmosEmulatorConnectionString(connectionString)) {
+        // Prevents self signed certificate error for emulator https://github.com/microsoft/vscode-cosmosdb/issues/1241#issuecomment-614446198
+        options.tlsAllowInvalidCertificates = true;
+    }
+
     try {
         return await MongoClient.connect(connectionString, options);
     } catch (err) {
         // Note: This file can't use `parseError` from `vscode-azureextensionui` because it's used by languageService.ts - see that file for more info
         const error = <{ message?: string, name?: string }>err;
-        const name = error && error.name;
         const message = error && error.message;
 
         // Example error: "failed to connect to server [localhost:10255] on first connect [MongoError: connect ECONNREFUSED 127.0.0.1:10255]"
         // Example error: "failed to connect to server [127.0.0.1:27017] on first connect [MongoError: connect ECONNREFUSED 127.0.0.1:27017]"
-        if (message && name === 'MongoError' && /ECONNREFUSED/.test(message) && /(localhost|127\.0\.0\.1)/.test(message)) {
-            throw new Error(`Unable to connect to local Mongo DB instance. Make sure it is started correctly. See ${Links.LocalConnectionDebuggingTips} for tips.\n${message}`);
+        if (message && /ECONNREFUSED/.test(message) && /(localhost|127\.0\.0\.1)/.test(message)) {
+            throw new MongoConnectError();
         }
 
         throw error;
     }
+}
+
+export class MongoConnectError extends Error {
+    constructor() {
+        super(`Unable to connect to local Mongo DB instance. Make sure it is started correctly. See ${Links.LocalConnectionDebuggingTips} for tips.`);
+    }
+}
+
+export function isCosmosEmulatorConnectionString(connectionString: string): boolean {
+    return connectionString.includes(encodeURIComponent(emulatorPassword));
 }