+++ src/mongo/registerMongoCommands.ts
@@ -4,13 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
-import { AzureTreeItem, callWithTelemetryAndErrorHandling, IActionContext, registerCommand, registerEvent } from "vscode-azureextensionui";
+import { AzureTreeItem, callWithTelemetryAndErrorHandling, IActionContext, IErrorHandlerContext, ITreeItemPickerContext, registerCommand, registerErrorHandler, registerEvent } from "vscode-azureextensionui";
+import { Experience, MongoExperience } from '../AzureDBExperiences';
 import { ext } from "../extensionVariables";
 import { AttachedAccountSuffix } from '../tree/AttachedAccountsTreeItem';
 import * as vscodeUtil from '../utils/vscodeUtils';
-import { MongoCollectionNodeEditor } from "./editors/MongoCollectionNodeEditor";
+import { MongoConnectError } from './connectToMongoClient';
 import { MongoDBLanguageClient } from "./languageClient";
-import { executeAllCommandsFromActiveEditor, executeCommandFromActiveEditor, executeCommandFromText, getAllErrorsFromTextDocument } from "./MongoScrapbook";
+import { executeAllCommandsFromActiveEditor, executeCommandFromActiveEditor, getAllErrorsFromTextDocument } from "./MongoScrapbook";
 import { MongoCodeLensProvider } from "./services/MongoCodeLensProvider";
 import { setConnectedNode } from "./setConnectedNode";
 import { MongoAccountTreeItem } from "./tree/MongoAccountTreeItem";
@@ -18,7 +19,7 @@ import { MongoCollectionTreeItem } from "./tree/MongoCollectionTreeItem";
 import { MongoDatabaseTreeItem } from "./tree/MongoDatabaseTreeItem";
 import { MongoDocumentTreeItem } from "./tree/MongoDocumentTreeItem";
 
-const connectedDBKey: string = 'ms-azuretools.vscode-cosmosdb.connectedDB';
+const connectedMongoKey: string = 'ms-azuretools.vscode-cosmosdb.connectedDB';
 let diagnosticsCollection: vscode.DiagnosticCollection;
 const mongoLanguageId: string = 'mongo';
 
@@ -41,10 +42,7 @@ export function registerMongoCommands(): MongoCodeLensProvider {
             node = <MongoAccountTreeItem>await ext.tree.showTreeItemPicker([MongoAccountTreeItem.contextValue, MongoAccountTreeItem.contextValue + AttachedAccountSuffix], context);
         }
         const databaseNode = <MongoDatabaseTreeItem>await node.createChild(context);
-        // reveal the database treeItem in case user cancels collection creation
-        await ext.treeView.reveal(databaseNode, { focus: false });
-        const collectionNode = <MongoCollectionTreeItem>await databaseNode.createChild(context);
-        await ext.treeView.reveal(collectionNode, { focus: true });
+        await databaseNode.createChild(context);
 
         await vscode.commands.executeCommand('cosmosDB.connectMongoDB', databaseNode);
     });
@@ -53,7 +51,6 @@ export function registerMongoCommands(): MongoCodeLensProvider {
             node = <MongoDatabaseTreeItem>await ext.tree.showTreeItemPicker(MongoDatabaseTreeItem.contextValue, context);
         }
         const collectionNode = await node.createChild(context);
-        await ext.treeView.reveal(collectionNode);
         await vscode.commands.executeCommand('cosmosDB.connectMongoDB', collectionNode.parent);
     });
     registerCommand('cosmosDB.createMongoDocument', async (context: IActionContext, node?: MongoCollectionTreeItem) => {
@@ -61,17 +58,18 @@ export function registerMongoCommands(): MongoCodeLensProvider {
             node = <MongoCollectionTreeItem>await ext.tree.showTreeItemPicker(MongoCollectionTreeItem.contextValue, context);
         }
         const documentNode = await node.createChild(context);
-        await ext.treeView.reveal(documentNode);
         await vscode.commands.executeCommand("cosmosDB.openDocument", documentNode);
     });
     registerCommand('cosmosDB.connectMongoDB', async (context: IActionContext, node?: MongoDatabaseTreeItem) => {
         if (!node) {
-            node = <MongoDatabaseTreeItem>await ext.tree.showTreeItemPicker(MongoDatabaseTreeItem.contextValue, context);
+            // Include defaultExperience in the context to prevent https://github.com/microsoft/vscode-cosmosdb/issues/1517
+            const experienceContext: ITreeItemPickerContext & { defaultExperience?: Experience } = { ...context, defaultExperience: MongoExperience };
+            node = <MongoDatabaseTreeItem>await ext.tree.showTreeItemPicker(MongoDatabaseTreeItem.contextValue, experienceContext);
         }
 
         const oldNodeId: string | undefined = ext.connectedMongoDB && ext.connectedMongoDB.fullId;
         await languageClient.connect(node.connectionString, node.databaseName);
-        ext.context.globalState.update(connectedDBKey, node.fullId);
+        ext.context.globalState.update(connectedMongoKey, node.fullId);
         setConnectedNode(node, codeLensProvider);
         await node.refresh();
 
@@ -84,23 +82,29 @@ export function registerMongoCommands(): MongoCodeLensProvider {
         }
     });
     registerCommand('cosmosDB.deleteMongoDB', async (context: IActionContext, node?: MongoDatabaseTreeItem) => {
+        const suppressCreateContext: ITreeItemPickerContext = context;
+        suppressCreateContext.suppressCreatePick = true;
         if (!node) {
             node = <MongoDatabaseTreeItem>await ext.tree.showTreeItemPicker(MongoDatabaseTreeItem.contextValue, context);
         }
         await node.deleteTreeItem(context);
         if (ext.connectedMongoDB && ext.connectedMongoDB.fullId === node.fullId) {
             setConnectedNode(undefined, codeLensProvider);
-            ext.context.globalState.update(connectedDBKey, undefined);
+            ext.context.globalState.update(connectedMongoKey, undefined);
             languageClient.disconnect();
         }
     });
     registerCommand('cosmosDB.deleteMongoCollection', async (context: IActionContext, node?: MongoCollectionTreeItem) => {
+        const suppressCreateContext: ITreeItemPickerContext = context;
+        suppressCreateContext.suppressCreatePick = true;
         if (!node) {
             node = <MongoCollectionTreeItem>await ext.tree.showTreeItemPicker(MongoCollectionTreeItem.contextValue, context);
         }
         await node.deleteTreeItem(context);
     });
     registerCommand('cosmosDB.deleteMongoDocument', async (context: IActionContext, node?: MongoDocumentTreeItem) => {
+        const suppressCreateContext: ITreeItemPickerContext = context;
+        suppressCreateContext.suppressCreatePick = true;
         if (!node) {
             node = <MongoDocumentTreeItem>await ext.tree.showTreeItemPicker(MongoDocumentTreeItem.contextValue, context);
         }
@@ -110,17 +114,13 @@ export function registerMongoCommands(): MongoCodeLensProvider {
         if (!node) {
             node = <MongoCollectionTreeItem>await ext.tree.showTreeItemPicker(MongoCollectionTreeItem.contextValue, context);
         }
-        await ext.editorManager.showDocument(context, new MongoCollectionNodeEditor(node), node.label + '-cosmos-collection.json');
+        await ext.fileSystem.showTextDocument(node);
     });
     registerCommand('cosmosDB.launchMongoShell', launchMongoShell);
     registerCommand('cosmosDB.newMongoScrapbook', async () => await vscodeUtil.showNewFile('', 'Scrapbook', '.mongo'));
-    registerCommand('cosmosDB.executeMongoCommand', async (context: IActionContext, commandText: object) => {
+    registerCommand('cosmosDB.executeMongoCommand', async (context: IActionContext, position?: vscode.Position) => {
         await loadPersistedMongoDBTask;
-        if (typeof commandText === "string") {
-            await executeCommandFromText(context, <string>commandText);
-        } else {
-            await executeCommandFromActiveEditor(context);
-        }
+        await executeCommandFromActiveEditor(context, position);
     });
     registerCommand('cosmosDB.executeAllMongoCommands', async (context: IActionContext) => {
         await loadPersistedMongoDBTask;
@@ -137,7 +137,7 @@ async function loadPersistedMongoDB(languageClient: MongoDBLanguageClient, codeL
         context.telemetry.properties.isActivationEvent = 'true';
 
         try {
-            const persistedNodeId: string | undefined = ext.context.globalState.get(connectedDBKey);
+            const persistedNodeId: string | undefined = ext.context.globalState.get(connectedMongoKey);
             if (persistedNodeId) {
                 const persistedNode = await ext.tree.findTreeItem(persistedNodeId, context);
                 if (persistedNode) {
@@ -191,6 +191,12 @@ function setUpErrorReporting(): void {
                 context.telemetry.suppressIfSuccessful = true;
             }
         });
+
+    registerErrorHandler((context: IErrorHandlerContext) => {
+        if (context.error instanceof MongoConnectError) {
+            context.errorHandling.suppressReportIssue = true;
+        }
+    });
 }
 
 function updateErrorsInScrapbook(context: IActionContext, document: vscode.TextDocument | undefined): void {