+++ src/postgres/commands/deletePostgresFunction.ts
@@ -4,18 +4,22 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { window } from 'vscode';
-import { DialogResponses, IActionContext } from "vscode-azureextensionui";
+import { DialogResponses, IActionContext, ITreeItemPickerContext } from "vscode-azureextensionui";
 import { ext } from "../../extensionVariables";
 import { localize } from "../../utils/localize";
 import { PostgresFunctionTreeItem } from "../tree/PostgresFunctionTreeItem";
 
 export async function deletePostgresFunction(context: IActionContext, treeItem?: PostgresFunctionTreeItem): Promise<void> {
+    const suppressCreateContext: ITreeItemPickerContext = context;
+    suppressCreateContext.suppressCreatePick = true;
     if (!treeItem) {
         treeItem = <PostgresFunctionTreeItem>await ext.tree.showTreeItemPicker(PostgresFunctionTreeItem.contextValue, { ...context, suppressCreatePick: true });
     }
 
-    const message = localize('deleteFunction', 'Are you sure you want to delete function "{0}"?', treeItem.label);
+    const message: string = localize('deleteFunction', 'Are you sure you want to delete function "{0}"?', treeItem.label);
     await ext.ui.showWarningMessage(message, { modal: true }, DialogResponses.deleteResponse);
     await treeItem.deleteTreeItem(context);
-    window.showInformationMessage(localize('successfullyDeletedFunction', 'Successfully deleted function "{0}".', treeItem.label));
+    const deleteMessage: string = localize('successfullyDeletedFunction', 'Successfully deleted function "{0}".', treeItem.label);
+    window.showInformationMessage(deleteMessage);
+    ext.outputChannel.appendLog(deleteMessage);
 }