+++ src/postgres/commands/openPostgresFunction.ts
@@ -5,7 +5,7 @@
 
 import { IActionContext } from "vscode-azureextensionui";
 import { ext } from "../../extensionVariables";
-import { PostgresFunctionEditor } from "../editors/PostgresFunctionEditor";
+import { showPostgresQuery } from "../showPostgresQuery";
 import { PostgresFunctionTreeItem } from "../tree/PostgresFunctionTreeItem";
 
 export async function openPostgresFunction(context: IActionContext, treeItem?: PostgresFunctionTreeItem): Promise<void> {
@@ -13,6 +13,5 @@ export async function openPostgresFunction(context: IActionContext, treeItem?: P
         treeItem = <PostgresFunctionTreeItem>await ext.tree.showTreeItemPicker(PostgresFunctionTreeItem.contextValue, context);
     }
 
-    const fileName: string = `${treeItem.label} (${treeItem.parent.parent.parent.server.name}.${treeItem.schema}).sql`;
-    await ext.editorManager.showDocument(context, new PostgresFunctionEditor(treeItem), fileName);
+    await showPostgresQuery(treeItem);
 }