+++ src/postgres/tree/PostgresFunctionTreeItem.ts
@@ -3,10 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Client } from "pg";
 import { AzureTreeItem, ISubscriptionContext, TreeItemIconPath } from "vscode-azureextensionui";
-import { getThemeAgnosticIconPath } from "../../constants";
-import { IPostgresProceduresQueryRow } from "../IPostgresProceduresQueryRow";
+import { getThemedIconPath } from "../../constants";
+import { IPostgresProceduresQueryRow } from "../getPostgresProcedureQueryRows";
+import { runPostgresQuery } from "../runPostgresQuery";
 import { PostgresFunctionsTreeItem } from "./PostgresFunctionsTreeItem";
 
 export class PostgresFunctionTreeItem extends AzureTreeItem<ISubscriptionContext> {
@@ -17,6 +17,7 @@ export class PostgresFunctionTreeItem extends AzureTreeItem<ISubscriptionContext
     public readonly schema: string;
     public readonly name: string;
     public readonly id: string;
+    public readonly args: string;
     public readonly isDuplicate: boolean;
     public definition: string;
 
@@ -25,6 +26,7 @@ export class PostgresFunctionTreeItem extends AzureTreeItem<ISubscriptionContext
         this.schema = row.schema;
         this.name = row.name;
         this.id = String(row.oid);
+        this.args = row.args;
         this.definition = row.definition;
         this.isDuplicate = isDuplicate;
     }
@@ -38,12 +40,10 @@ export class PostgresFunctionTreeItem extends AzureTreeItem<ISubscriptionContext
     }
 
     public get iconPath(): TreeItemIconPath {
-        return getThemeAgnosticIconPath('Collection.svg');
+        return getThemedIconPath('function.svg');
     }
 
     public async deleteTreeItemImpl(): Promise<void> {
-        const client = new Client(this.parent.clientConfig);
-        await client.connect();
-        await client.query(`DROP FUNCTION ${this.schema}.${this.name};`);
+        await runPostgresQuery(this.parent.clientConfig, `DROP FUNCTION ${this.schema}.${this.name}(${this.args});`);
     }
 }