+++ src/postgres/tree/PostgresServerTreeItem.ts
@@ -16,6 +16,8 @@ import { nonNullProp } from '../../utils/nonNull';
 import { PostgresDatabaseTreeItem } from './PostgresDatabaseTreeItem';
 import { PostgresFunctionsTreeItem } from './PostgresFunctionsTreeItem';
 import { PostgresFunctionTreeItem } from './PostgresFunctionTreeItem';
+import { PostgresStoredProceduresTreeItem } from './PostgresStoredProceduresTreeItem';
+import { PostgresStoredProcedureTreeItem } from './PostgresStoredProcedureTreeItem';
 import { PostgresTablesTreeItem } from './PostgresTablesTreeItem';
 import { PostgresTableTreeItem } from './PostgresTableTreeItem';
 
@@ -85,6 +87,8 @@ export class PostgresServerTreeItem extends AzureParentTreeItem<ISubscriptionCon
             case PostgresTableTreeItem.contextValue:
             case PostgresFunctionsTreeItem.contextValue:
             case PostgresFunctionTreeItem.contextValue:
+            case PostgresStoredProceduresTreeItem.contextValue:
+            case PostgresStoredProcedureTreeItem.contextValue:
                 return true;
             default:
                 return false;
@@ -110,6 +114,7 @@ export class PostgresServerTreeItem extends AzureParentTreeItem<ISubscriptionCon
         const deletingMessage: string = `Deleting server "${this.name}"...`;
         await vscode.window.withProgress({ location: vscode.ProgressLocation.Notification, title: deletingMessage }, async () => {
             await client.servers.deleteMethod(this.resourceGroup, this.name);
+            await this.deletePostgresCredentials();
         });
     }
 
@@ -136,6 +141,20 @@ export class PostgresServerTreeItem extends AzureParentTreeItem<ISubscriptionCon
         // `semver.gte` complains when a version doesn't have decimals (i.e. "10"), so attempt to convert version to SemVer
         const version: SemVer | null = coerce(this.server.version);
         return !!version && gte(version, '11.0.0');
+    }
+
+    private async deletePostgresCredentials(): Promise<void> {
+        if (ext.keytar) {
+            const serviceName: string = PostgresServerTreeItem.serviceName;
+            const storedValue: string | undefined = ext.context.globalState.get(serviceName);
+            let servers: IPersistedServer[] = storedValue ? JSON.parse(storedValue) : [];
+
+            // Remove this server from the cache
+            servers = servers.filter((server: IPersistedServer) => { return server.id !== this.id; });
+
+            await ext.context.globalState.update(serviceName, JSON.stringify(servers));
+            await ext.keytar.deletePassword(serviceName, this.id);
+        }
     }
 }